Refactored ReplyFlow to remove broken modularization design smell
